Remove useless task subclass.
