+++ src/main/java/seedu/parser/Parser.java
@@ -1,17 +1,8 @@
 package seedu.parser;
 
-import seedu.command.Command;
-import seedu.command.EditCommand;
-import seedu.command.ListCommand;
-import seedu.command.DeadlineCommand;
-import seedu.command.DeleteCommand;
-import seedu.command.EventCommand;
-import seedu.command.ExitCommand;
-import seedu.command.FindCommand;
-import seedu.command.HelpCommand;
-import seedu.command.TodoCommand;
-import seedu.command.ClearCommand;
+import seedu.command.*;
 
+import static seedu.common.Constants.ADD;
 import static seedu.common.Constants.CLEAR;
 import static seedu.common.Constants.FIND;
 import static seedu.common.Constants.EVENT;
@@ -44,14 +35,8 @@ public class Parser {
             // return prepareDoneCommand(commandSections[1], wordLength);
         case HELP:
             return new HelpCommand();
-        case DELETE:
-            return new DeleteCommand(userCommand);
-        case TODO:
-            return new TodoCommand(userCommand);
-        case EVENT:
-            return new EventCommand(userCommand);
-        case DEADLINE:
-            return new DeadlineCommand(userCommand);
+        case ADD:
+            return new AddCommand(userCommand);
         case FIND:
             return new FindCommand(userCommand);
         case EDIT:
