--- src/main/java/seedu/tasks/Deadline.java
@@ -1,24 +1,0 @@
-package seedu.tasks;
-
-import java.text.DateFormat;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.Date;
-import java.util.Locale;
-
-public class Deadline extends Task {
-    /**
-     * Initializes Deadline task.
-     *
-     * @param title title of deadline if any.
-     * @param description description of deadline if any.
-     * @param date date in format:yyyy-mm-dd of deadline if any.
-     * @param time time in format: hh:mm of deadline if any.
-     * @param location location of deadline if any.
-     * @param reminder reminder of deadline if any.
-     */
-    public Deadline(String title, String description, String date,String time, String location, String reminder) {
-        super(title, description,date,time,location,reminder);
-    }
-
-}
